fixed: routing issue and implemented accordion in transactions
