+++ app/(authenticated)/(tabs)/_layout.tsx
@@ -1,10 +1,8 @@
 import { Tabs } from "expo-router";
 import React from "react";
-import HomeHeader from "~/components/HomeHeader";
 import colors from "~/lib/colors";
 import { Home } from "~/lib/icons/Home";
 import { Settings } from "~/lib/icons/Settings";
-import { Send } from "~/lib/icons/Send";
 
 export default function TabLayout() {
 	return (
@@ -17,15 +15,15 @@ export default function TabLayout() {
 				name="home"
 				options={{
 					title: "Home",
+					headerShown: false,
 					tabBarIcon: ({ size, color }) => (
 						<Home strokeWidth={1.25} size={size} color={color} />
 					),
-					header: () => <HomeHeader />,
 				}}
 			/>
 
 			<Tabs.Screen
-				name="send"
+				name="[contract]"
 				options={{
 					headerShown: false,
 					href: null,
@@ -35,6 +33,7 @@ export default function TabLayout() {
 				name="settings"
 				options={{
 					title: "Settings",
+					headerShown: false,
 					tabBarIcon: ({ size, color }) => (
 						<Settings strokeWidth={1.25} size={size} color={color} />
 					),
