+++ package.json
@@ -35,19 +35,21 @@
     "@stacks/blockchain-api-client": "^8.4.0",
     "@stacks/transactions": "^7.0.2",
     "@stacks/wallet-sdk": "^7.0.2",
+    "accordion-collapse-react-native": "^1.1.1",
     "axios": "^1.7.9",
     "buffer": "^6.0.3",
+    "caniuse-lite": "^1.0.30001690",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "crypto-browserify": "^3.12.1",
-    "expo": "^52.0.21",
+    "expo": "^52.0.23",
     "expo-asset": "~11.0.1",
     "expo-blur": "^14.0.1",
     "expo-clipboard": "~7.0.0",
     "expo-image": "^2.0.3",
     "expo-linking": "~7.0.3",
     "expo-navigation-bar": "~4.0.6",
-    "expo-router": "~4.0.14",
+    "expo-router": "~4.0.15",
     "expo-secure-store": "^14.0.0",
     "expo-splash-screen": "~0.29.18",
     "expo-status-bar": "~2.0.0",
@@ -63,6 +65,7 @@
     "react-native": "0.76.5",
     "react-native-actions-sheet": "^0.9.7",
     "react-native-blurhash": "^2.1.0",
+    "react-native-collapsible": "^1.6.2",
     "react-native-gesture-handler": "^2.21.2",
     "react-native-get-random-values": "^1.11.0",
     "react-native-mmkv": "^3.2.0",
@@ -73,7 +76,7 @@
     "react-native-web": "~0.19.13",
     "react-query": "^3.39.3",
     "readable-stream": "^4.6.0",
-    "tailwind-merge": "^2.5.5",
+    "tailwind-merge": "^2.6.0",
     "tailwindcss": "3.3.5",
     "tailwindcss-animate": "^1.0.7",
     "text-encoding": "^0.7.0",
